Increase c_stdlib version to 2.34 to get kernel 5.14 headers
systemd v258 increased the minimum baseline kernel version to 5.4,
making it increasingly difficult to maintain patches that get this to
build with c_stdlib 2.17 (kernel 3.10) or even c_stdlib 2.28 (kernel
4.18). So move to c_stdlib 2.34 which brings kernel 5.14 headers which
is higher than the new baseline version (5.4) and even the recommended
baseline (5.7). We will maintain a branch with systemd v257 to ensure
older systems are still usable.